requirements.txt
fastapi
uvicorn
pydantic
pandas
requests
mysql-connector-python